[refactor] Add IDGenerator and replace makeRandomID()
- IDGenerator 추가 및 관련 코드 대체
